(PC-35943)[API] chore: move create_or_update_ean_offer job to offers module
